% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_feature_selector.R
\name{.evaluate_split_score}
\alias{.evaluate_split_score}
\title{Evaluate Split Score}
\usage{
.evaluate_split_score(target, f, sp, criteria_type, ig_metric)
}
\arguments{
\item{target}{Vector of class labels.}

\item{f}{Feature vector to split on.}

\item{sp}{Split point (numeric threshold or factor level).}

\item{criteria_type}{Splitting criterion: "gini", "info_gain", or "gain_ratio".}

\item{ig_metric}{Impurity metric to use if criteria_type is "info_gain": "gini" or "entropy".}
}
\value{
Numeric score of the split (higher is better).
}
\description{
Evaluate the quality of a candidate binary split using a chosen criterion.
}
\keyword{internal}
