% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_feature_selector.R
\name{gain_ratio}
\alias{gain_ratio}
\title{Gain Ratio}
\usage{
gain_ratio(target, feature_factor)
}
\arguments{
\item{target}{Vector of class labels.}

\item{feature_factor}{Factor variable used for split.}
}
\value{
Numeric value of gain ratio.
}
\description{
Compute gain ratio for a factor variable split.
}
\keyword{internal}
