% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_feature_selector.R
\name{info_gain}
\alias{info_gain}
\title{Information Gain}
\usage{
info_gain(feature, target, metric = c("entropy", "gini"))
}
\arguments{
\item{feature}{Factor variable used to split data.}

\item{target}{Vector of class labels.}

\item{metric}{Choice of impurity metric: "entropy" or "gini".}
}
\value{
Numeric value of information gain.
}
\description{
Compute information gain from splitting on a factor variable.
}
\keyword{internal}
