% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stree_fit_svm.R
\name{stree_fit_binary_svm}
\alias{stree_fit_binary_svm}
\title{Fit Binary SVM for STree}
\usage{
stree_fit_binary_svm(X, y, kernel, verbose = FALSE, ...)
}
\description{
Fit Binary SVM for STree
}
\keyword{internal}
